requirements.txt
Django==5.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1